services/tokenization/requirements.txt
fastapi
uvicorn
pydantic-settings
grpcio
protobuf
sqlalchemy
asyncpg
python-jose[cryptography]
